+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
  
 - new method for get the name of the current method
 - new method for get the name of the calling method
+- new method for get H2 connection created
 
 CHANGED: